fix: switched to use int as PK instead of UUID due to inscrutable SQLalchemy error (as below).
sqlalchemy.exc.InvalidRequestError: Can't match sentinel values in result set to parameter sets; key '0f694c62-7627-471c-ba12-f382a14fa985' was not found. There may be a mismatch between the datatype passed to the DBAPI driver vs. that which it returns in a result row.  Ensure the given Python value matches the expected result type *exactly*, taking care to not rely upon implicit conversions which may occur such as when using strings in place of UUID or integer values, etc.